+++ src/headers/src/RunTimer.cpp
@@ -6,6 +6,8 @@
 
 #include <Geode/Geode.hpp>
 
+#include <Geode/utils/terminate.hpp>
+
 using namespace geode::prelude;
 
 #ifndef GEODE_IS_IOS
@@ -13,33 +15,63 @@ using namespace geode::prelude;
 using namespace keybinds;
 #endif
 
+class RunTimer::Impl final {
+public:
+    Mod* m_srtMod = getMod(); // it's modding time :3
+
+    ccColor3B m_col = m_srtMod->getSettingValue<ccColor3B>("color"); // The color of the speedrun timer
+    ccColor3B m_colPause = m_srtMod->getSettingValue<ccColor3B>("color-pause"); // The color of the speedrun timer when paused
+    ccColor3B m_colStart = m_srtMod->getSettingValue<ccColor3B>("color-start"); // The color of the speedrun timer before starting
+
+    CCScheduler* m_scheduler = nullptr; // The scheduler for the speedrun timer
+
+    float m_runTime = 0.f; // Current time
+    float m_lastSplitTime = 0.f; // Last time a split was created
+
+    bool m_speedtimerOn = false; // If the speedrun is active
+    bool m_speedtimerPaused = true; // If the speedrun is paused
+
+    CCMenu* m_timeMenu = nullptr; // The menu that contains the timers
+
+    CCLabelBMFont* m_speedtimer = nullptr; // The text label for the time
+    CCLabelBMFont* m_speedtimerMs = nullptr; // The text label for the time in milliseconds
+
+    ScrollLayer* m_splitList = nullptr; // The scrolling list of timer splits
+};
+
+RunTimer::RunTimer() {
+    m_impl = std::make_unique<Impl>();
+};
+
+RunTimer::~RunTimer() {};
+
 bool RunTimer::init() {
-    m_scheduler = CCScheduler::get();
+    m_impl->m_scheduler = CCScheduler::get();
 
     if (CCNode::init()) {
         setID("timer"_spr);
         setContentSize({ 125.f, 37.5f });
 
         // assign the milliseconds timer
-        m_speedtimerMs = CCLabelBMFont::create(".00", "gjFont16.fnt");
-        m_speedtimerMs->setID("timer-milliseconds");
-        m_speedtimerMs->setScale(0.5f);
-        m_speedtimerMs->setColor(m_colStart);
-        m_speedtimerMs->setAlignment(CCTextAlignment::kCCTextAlignmentLeft);
-        m_speedtimerMs->setPosition({ getScaledContentWidth() - (m_speedtimerMs->getScaledContentWidth() + 2.5f), 0.5f });
-        m_speedtimerMs->setAnchorPoint({ 0, 0 });
+        m_impl->m_speedtimerMs = CCLabelBMFont::create(".00", "gjFont16.fnt");
+        m_impl->m_speedtimerMs->setID("timer-milliseconds");
+        m_impl->m_speedtimerMs->setScale(0.5f);
+        m_impl->m_speedtimerMs->setColor(m_impl->m_colStart);
+        m_impl->m_speedtimerMs->setAlignment(CCTextAlignment::kCCTextAlignmentLeft);
+        m_impl->m_speedtimerMs->setPosition({ getScaledContentWidth() - (m_impl->m_speedtimerMs->getScaledContentWidth() + 2.5f), 0.5f });
+        m_impl->m_speedtimerMs->setAnchorPoint({ 0, 0 });
 
         // assign the seconds timer
-        m_speedtimer = CCLabelBMFont::create("0", "gjFont16.fnt");
-        m_speedtimer->setID("timer-seconds");
-        m_speedtimer->setScale(0.875f);
-        m_speedtimer->setColor(m_colStart);
-        m_speedtimer->setAlignment(CCTextAlignment::kCCTextAlignmentRight);
-        m_speedtimer->setPosition({ getScaledContentWidth() - (m_speedtimerMs->getScaledContentWidth() + 3.75f), 0.5f });
-        m_speedtimer->setAnchorPoint({ 1, 0 });
+        m_impl->m_speedtimer = CCLabelBMFont::create("0", "gjFont16.fnt");
+        m_impl->m_speedtimer->setID("timer-seconds");
+        m_impl->m_speedtimer->setScale(0.875f);
+        m_impl->m_speedtimer->setColor(m_impl->m_colStart);
+        m_impl->m_speedtimer->setAlignment(CCTextAlignment::kCCTextAlignmentRight);
+        m_impl->m_speedtimer->setPosition({ getScaledContentWidth() - (m_impl->m_speedtimerMs->getScaledContentWidth() + 3.75f), 0.5f });
+        m_impl->m_speedtimer->setAnchorPoint({ 1, 0 });
 
-        addChild(m_speedtimerMs);
-        addChild(m_speedtimer);
+        addChild(m_impl->m_speedtimerMs);
+        addChild(m_impl->m_speedtimer);
 
         // Create layout for scroll layer
         auto scrollLayerLayout = ColumnLayout::create()
@@ -62,17 +94,16 @@ bool RunTimer::init() {
             scroll->m_contentLayer->setLayout(scrollLayerLayout);
             scroll->m_contentLayer->setTouchEnabled(false);
 
-            m_splitList = scroll;
+            m_impl->m_splitList = scroll;
+            addChild(m_impl->m_splitList);
 
-            addChild(m_splitList);
-
-            m_splitList->m_contentLayer->updateLayout(true);
-            m_splitList->scrollToTop();
+            m_impl->m_splitList->m_contentLayer->updateLayout(true);
+            m_impl->m_splitList->scrollToTop();
         } else {
             log::error("Failed to create scroll layer for speedrun splits");
         };
 
-        auto bgOpacity = static_cast<int>(m_srtMod->getSettingValue<int64_t>("bg-opacity"));
+        auto bgOpacity = static_cast<int>(m_impl->m_srtMod->getSettingValue<int64_t>("bg-opacity"));
 
         auto bg = CCLayerColor::create({ 0, 0, 0, 255 });
         bg->setID("background");
@@ -87,8 +118,8 @@ bool RunTimer::init() {
 #ifndef GEODE_IS_IOS
         // toggle timer
         this->template addEventListener<InvokeBindFilter>([=](InvokeBindEvent* event) {
-            if (event->isDown()) pauseTimer(!m_speedtimerPaused); // toggle the timer on or off
-            log::info("Speedrun timer set to {}", m_speedtimerPaused ? "paused" : "resumed");
+            if (event->isDown()) pauseTimer(!m_impl->m_speedtimerPaused); // toggle the timer on or off
+            log::info("Speedrun timer set to {}", m_impl->m_speedtimerPaused ? "paused" : "resumed");
 
             return ListenerResult::Propagate;
                                                           }, "pause-timer"_spr);
@@ -96,7 +127,7 @@ bool RunTimer::init() {
         // create a split
         this->template addEventListener<InvokeBindFilter>([=](InvokeBindEvent* event) {
             if (event->isDown()) createSplit(); // create a split at the current time
-            log::info("Speedrun split created at {} seconds", m_runTime);
+            log::info("Speedrun split created at {} seconds", m_impl->m_runTime);
 
             return ListenerResult::Propagate;
                                                           }, "split-timer"_spr);
@@ -110,9 +141,9 @@ bool RunTimer::init() {
                                                           }, "reset-timer"_spr);
 #endif
 
-        setScale(static_cast<float>(m_srtMod->getSettingValue<double>("scale")));
+        setScale(static_cast<float>(m_impl->m_srtMod->getSettingValue<double>("scale")));
 
-        if (m_scheduler) m_scheduler->scheduleUpdateForTarget(this, 0, false);
+        if (m_impl->m_scheduler) m_impl->m_scheduler->scheduleUpdateForTarget(this, 0, false);
 
         return true;
     } else {
@@ -122,16 +153,16 @@ bool RunTimer::init() {
 
 void RunTimer::update(float dt) {
     // always go at 1x speed
-    auto dtScale = m_scheduler->getTimeScale();
+    auto dtScale = m_impl->m_scheduler->getTimeScale();
     if (dtScale != 0.f) dt = dt / dtScale;
 
-    if (m_speedtimerOn) m_runTime += m_speedtimerPaused ? 0.f : dt;
+    if (m_impl->m_speedtimerOn) m_impl->m_runTime += m_impl->m_speedtimerPaused ? 0.f : dt;
 
-    auto newTime = static_cast<int>(m_runTime);
+    auto newTime = static_cast<int>(m_impl->m_runTime);
     std::string secStr = std::to_string(newTime);
 
-    if (m_srtMod->getSettingValue<bool>("format-minutes")) {
-        int minutes = static_cast<int>(m_runTime / 60.f);
+    if (m_impl->m_srtMod->getSettingValue<bool>("format-minutes")) {
+        int minutes = static_cast<int>(m_impl->m_runTime / 60.f);
         int seconds = newTime % 60;
 
         if (minutes > 0) {
@@ -145,54 +176,54 @@ void RunTimer::update(float dt) {
         secStr = std::to_string(newTime);
     };
 
-    if (m_speedtimer) m_speedtimer->setCString(secStr.c_str()); // seconds
+    if (m_impl->m_speedtimer) m_impl->m_speedtimer->setCString(secStr.c_str()); // seconds
 
-    if (m_speedtimerMs) { // ms
-        int ms = static_cast<int>(m_runTime * 100) % 100;
+    if (m_impl->m_speedtimerMs) { // ms
+        int ms = static_cast<int>(m_impl->m_runTime * 100) % 100;
 
         std::ostringstream oss;
         oss << "." << (ms > 9 ? "" : "0") << ms;
 
-        m_speedtimerMs->setCString(oss.str().c_str());
+        m_impl->m_speedtimerMs->setCString(oss.str().c_str());
     };
 };
 
 void RunTimer::toggleTimer(bool toggle) {
-    if (m_speedtimerOn == toggle) {
+    if (m_impl->m_speedtimerOn == toggle) {
         log::info("Speedrun timer is already {}", toggle ? "on" : "off");
     } else {
-        m_speedtimerOn = toggle;
+        m_impl->m_speedtimerOn = toggle;
 
-        if (m_speedtimerOn) {
-            auto color = m_runTime < 0.f ? m_col : m_colStart;
+        if (m_impl->m_speedtimerOn) {
+            auto color = m_impl->m_runTime < 0.f ? m_impl->m_col : m_impl->m_colStart;
 
-            if (m_speedtimer) m_speedtimer->setColor(color);
-            if (m_speedtimerMs) m_speedtimerMs->setColor(color);
+            if (m_impl->m_speedtimer) m_impl->m_speedtimer->setColor(color);
+            if (m_impl->m_speedtimerMs) m_impl->m_speedtimerMs->setColor(color);
 
             log::info("Speedrun timer started");
         } else {
-            if (m_speedtimer) m_speedtimer->setColor(m_colStart);
-            if (m_speedtimerMs) m_speedtimerMs->setColor(m_colStart);
+            if (m_impl->m_speedtimer) m_impl->m_speedtimer->setColor(m_impl->m_colStart);
+            if (m_impl->m_speedtimerMs) m_impl->m_speedtimerMs->setColor(m_impl->m_colStart);
 
-            log::info("Speedrun timer stopped at {} seconds", m_runTime);
+            log::info("Speedrun timer stopped at {} seconds", m_impl->m_runTime);
         };
     };
 };
 
 void RunTimer::pauseTimer(bool pause) {
-    if (m_speedtimerPaused == pause) {
+    if (m_impl->m_speedtimerPaused == pause) {
         log::info("Speedrun timer is already {}", pause ? "paused" : "unpaused");
     } else {
-        m_speedtimerPaused = pause;
+        m_impl->m_speedtimerPaused = pause;
 
-        if (m_speedtimerPaused) {
-            if (m_speedtimer) m_speedtimer->setColor(m_colPause);
-            if (m_speedtimerMs) m_speedtimerMs->setColor(m_colPause);
+        if (m_impl->m_speedtimerPaused) {
+            if (m_impl->m_speedtimer) m_impl->m_speedtimer->setColor(m_impl->m_colPause);
+            if (m_impl->m_speedtimerMs) m_impl->m_speedtimerMs->setColor(m_impl->m_colPause);
 
-            log::info("Speedrun timer paused at {} seconds", m_runTime);
+            log::info("Speedrun timer paused at {} seconds", m_impl->m_runTime);
         } else {
-            if (m_speedtimer) m_speedtimer->setColor(m_col);
-            if (m_speedtimerMs) m_speedtimerMs->setColor(m_col);
+            if (m_impl->m_speedtimer) m_impl->m_speedtimer->setColor(m_impl->m_col);
+            if (m_impl->m_speedtimerMs) m_impl->m_speedtimerMs->setColor(m_impl->m_col);
 
             log::info("Speedrun timer unpaused");
         };
@@ -200,22 +231,22 @@ void RunTimer::pauseTimer(bool pause) {
 };
 
 void RunTimer::createSplit() {
-    if (m_speedtimerOn) {
-        auto splitDelta = m_runTime - m_lastSplitTime;
-        m_lastSplitTime = m_runTime;
+    if (m_impl->m_speedtimerOn) {
+        auto splitDelta = m_impl->m_runTime - m_impl->m_lastSplitTime;
+        m_impl->m_lastSplitTime = m_impl->m_runTime;
 
-        if (auto splitNode = SplitSegment::create(m_runTime, splitDelta)) {
+        if (auto splitNode = SplitSegment::create(m_impl->m_runTime, splitDelta)) {
 
-            auto limit = static_cast<int>(m_srtMod->getSettingValue<int64_t>("split-limit"));
-            auto withinLimit = (limit - 1) >= m_splitList->m_contentLayer->getChildrenCount();
+            auto limit = static_cast<int>(m_impl->m_srtMod->getSettingValue<int64_t>("split-limit"));
+            auto withinLimit = (limit - 1) >= m_impl->m_splitList->m_contentLayer->getChildrenCount();
 
-            if (m_splitList) {
+            if (m_impl->m_splitList) {
                 if (withinLimit) {
                     log::info("Adding split node to split list");
                 } else {
                     log::warn("Split limit reached, removing first split node");
 
-                    auto children = m_splitList->m_contentLayer->getChildren();
+                    auto children = m_impl->m_splitList->m_contentLayer->getChildren();
                     if (children && children->count() > 0) {
                         auto firstChild = static_cast<CCNode*>(children->objectAtIndex(0));
                         if (firstChild) firstChild->removeMeAndCleanup();
@@ -224,14 +255,14 @@ void RunTimer::createSplit() {
                     };
                 };
 
-                m_splitList->m_contentLayer->addChild(splitNode);
-                m_splitList->m_contentLayer->updateLayout(true);
+                m_impl->m_splitList->m_contentLayer->addChild(splitNode);
+                m_impl->m_splitList->m_contentLayer->updateLayout(true);
             } else {
                 splitNode->removeMeAndCleanup();
                 log::error("Split list not found");
             };
 
-            log::info("Speedrun split created at {} seconds{}", m_runTime, withinLimit ? "" : " (limit reached, node not added)");
+            log::info("Speedrun split created at {} seconds{}", m_impl->m_runTime, withinLimit ? "" : " (limit reached, node not added)");
         } else {
             log::error("Failed to create speedrun split node");
         };
@@ -241,28 +272,28 @@ void RunTimer::createSplit() {
 };
 
 void RunTimer::resetAll() {
-    m_runTime = 0.f;
-    m_lastSplitTime = 0.f;
+    m_impl->m_runTime = 0.f;
+    m_impl->m_lastSplitTime = 0.f;
 
-    m_speedtimerPaused = true;
+    m_impl->m_speedtimerPaused = true;
 
-    if (m_speedtimer) {
-        m_speedtimer->setCString("0");
-        m_speedtimer->setColor(m_colStart);
+    if (m_impl->m_speedtimer) {
+        m_impl->m_speedtimer->setCString("0");
+        m_impl->m_speedtimer->setColor(m_impl->m_colStart);
     } else {
         log::error("Speedrun timer label not found");
     };
 
-    if (m_speedtimerMs) {
-        m_speedtimerMs->setCString(".00");
-        m_speedtimerMs->setColor(m_colStart);
+    if (m_impl->m_speedtimerMs) {
+        m_impl->m_speedtimerMs->setCString(".00");
+        m_impl->m_speedtimerMs->setColor(m_impl->m_colStart);
     } else {
         log::error("Speedrun timer milliseconds label not found");
     };
 
-    if (m_splitList) {
-        m_splitList->m_contentLayer->removeAllChildren();
-        m_splitList->m_contentLayer->updateLayout(true);
+    if (m_impl->m_splitList) {
+        m_impl->m_splitList->m_contentLayer->removeAllChildren();
+        m_impl->m_splitList->m_contentLayer->updateLayout(true);
     } else {
         log::error("Split list not found");
     };
@@ -271,7 +302,7 @@ void RunTimer::resetAll() {
 };
 
 bool RunTimer::isTimerPaused() {
-    return m_speedtimerPaused;
+    return m_impl->m_speedtimerPaused;
 };
 
 RunTimer* RunTimer::create() {